add DoctorScopeFunc detection + README
this outlines in the README how to set up an example git root implementation
for this. it's a tad annoying that scope does not do any correcting for
absolute file paths that are within the $GOPATH, but it's not too hard to add
for now to get around this, briefly looked at adding that into core godoctor,
but has obvious ux implications so just did it here.

initially, i had just defined a command like `DoctorScopeFunc` that
sets `g:doctor_scope` instead of returning the desired output, and this works
of course, but it's nice to just define this once in my vim file and not have
to think about it and let the doctor order it up itself instead of having to
remember to call it before i do a refactoring.